review: Move saved review state out of terminal rendering
Default page selection now lives in data.file_review. An output module
still builds the review state saved between commands from selected lines,
pages, actions, and fingerprints used to detect file changes.

That saved state describes what the user reviewed and selected; it does not
render terminal content. Keeping the builder in output makes command
behavior depend on presentation code.

This commit moves saved review-state construction beside page selection in
data.file_review and removes the output-side builder.

The next commit will update behavior and architecture tests for the new
data-side builder.
