+++ src/git_stage_batch/commands/batch_source/file_display_action.py
@@ -21,6 +21,10 @@ from ...data.file_review.state import (
     clear_last_file_review_state,
     write_last_file_review_state,
 )
+from ...data.file_review.state_builder import (
+    make_file_review_state,
+    resolve_default_review_pages,
+)
 from ...data.selected_change.file_changes import (
     cache_binary_file_change,
     cache_gitlink_change,
@@ -32,8 +36,6 @@ from ...exceptions import exit_with_error
 from ...i18n import _
 from ...output.file_review import print_file_review
 from ...output.file_review_model_builder import build_file_review_model
-from ...data.file_review.state_builder import resolve_default_review_pages
-from ...output.file_review_state_builder import make_file_review_state
 from ...output.hunk import print_line_level_changes
 from ...output.patch import (
     print_binary_file_change,
